Add files via upload
This consist of Jupyter notebook file with ML code. There are some images of person with pneumonia and non-pneumonia and model is trained on such images. There is two set of data train and val dataset to train our ML model. Post training the model will predict pneumonia or not based on x-ray input image.